+++ src/assets/icons/calendar@3x.png
