+++ src/assets/icons/camera.png
