+++ src/assets/icons/clock@2x.png
