+++ src/assets/icons/restaurant@3x.png
